Rectify some input stuff.

+++ Source/iop/namco_sys147/Iop_NamcoSys147.h
@@ -86,7 +86,7 @@ namespace Iop
 			std::unique_ptr<Framework::CHttpServer> m_ioServer;
 			std::mutex m_barcodeMutex;
 			std::string m_currentBarcode;
-			uint32 m_systemSwitchState = ~0U;
+			uint16 m_systemSwitchState = ~0U;
 			uint16 m_playerSwitchState = ~0U;
 		};
 	}
